+++ src/test/java/de/muenchen/mobidam/TestConstants.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
+ * der Landeshauptstadt München, 2022
+ */
+package de.muenchen.mobidam;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class TestConstants {
+    public static final String SPRING_NO_SECURITY_PROFILE = "no-security";
+}