+++ RELEASENOTES.md
@@ -1,6 +1,10 @@
 # Release-Notes
 
-## Sprint 7 (05.03.2024 - 26.03.2024)
+## Sprint 9 (26.03.2024 - 16.04.2024)
+### Hinzugefügt
+- Security Absicherung
+
+## Sprint 8 (05.03.2024 - 26.03.2024)
 ### Hinzugefügt
 - Archivierung
 - Url vorsignierte Adresse erstellen